Module Core.Tests

val properties_to_test : loc:Ppxlib.location -> name:string -> ?⁠sig_item:Ppxlib.signature_item -> Properties.t -> Ppxlib.structure_item list

properties_to_tests loc name sig_item properties create a test for every property inside properties.

Returns the new list of tests as a structure_item list including QCheck.Test.t and add them to the test suite.