Module Ppx_pbt__.Helpers

val pbt_name : string

pbt_name constant name for attributes

val extract_name_from_pattern : Ppxlib.pattern -> string option

extract_name_from_pattern pat tries to extract the function name located in the pattern

let <pattern> = <expr> 

val filter_attributes : string -> Ppxlib.attribute list -> Ppxlib.attribute list

filter_attributes expected attributes filters attributes with name expected

val from_string : string -> Core.Properties.t

from_string properties parse properties and returns a Properties.t

val get_properties : Ppxlib.attribute list -> Core.Properties.property list

get_properties attributes returns the list propertiy inside attributes

Step 1: keep every attribute named pbt_name Step 3: extract each attribute's payload, which must be a string constant Step 3: parse the properties Step 4: concat every properties into a single list

Implicitly the function returns an empty list of properties if there is not properties attached on the attributes