Module Ppx_pbt__.Interface

val interface : string -> bool

interface file_name returns true if there is an environement for file_name

val inline_impl_tests : Ppxlib.structure -> Ppxlib.structure

inline_impl_tests str replaces the according specification in mli with the actual implementation inside str.

The file_name is fetched from a structure_item inside str and will use the previously stored environment according to the file_name

val intf : string -> Ppxlib.signature -> Ppxlib.signature

intf file_name signature fetchs properties attached to attributes inside each signature_item in signature. Creates an environment for file_name which would be later used to inline tests in implementation files